windows compatability

+++ src/RcppExports.cpp
@@ -6,7 +6,7 @@
 using namespace Rcpp;
 
 // Encrypt
-RawVector Encrypt(RawVector input, CharacterVector passwd, size_t maxmem = 0, double maxmemfrac = 0.1, double maxtime = 1.0);
+RawVector Encrypt(RawVector input, CharacterVector passwd, uint32_t maxmem = 0, double maxmemfrac = 0.1, double maxtime = 1.0);
 RcppExport SEXP scrypt_Encrypt(SEXP inputSEXP, SEXP passwdSEXP, SEXP maxmemSEXP, SEXP maxmemfracSEXP, SEXP maxtimeSEXP) {
 BEGIN_RCPP
     SEXP __sexp_result;
@@ -14,7 +14,7 @@ BEGIN_RCPP
         Rcpp::RNGScope __rngScope;
         Rcpp::traits::input_parameter< RawVector >::type input(inputSEXP );
         Rcpp::traits::input_parameter< CharacterVector >::type passwd(passwdSEXP );
-        Rcpp::traits::input_parameter< size_t >::type maxmem(maxmemSEXP );
+        Rcpp::traits::input_parameter< uint32_t >::type maxmem(maxmemSEXP );
         Rcpp::traits::input_parameter< double >::type maxmemfrac(maxmemfracSEXP );
         Rcpp::traits::input_parameter< double >::type maxtime(maxtimeSEXP );
         RawVector __result = Encrypt(input, passwd, maxmem, maxmemfrac, maxtime);
@@ -41,14 +41,14 @@ BEGIN_RCPP
 END_RCPP
 }
 // HashPassword
-CharacterVector HashPassword(CharacterVector passwd, size_t maxmem = 0, double maxmemfrac = 0.1, double maxtime = 1.0);
+CharacterVector HashPassword(CharacterVector passwd, uint32_t maxmem = 0, double maxmemfrac = 0.1, double maxtime = 1.0);
 RcppExport SEXP scrypt_HashPassword(SEXP passwdSEXP, SEXP maxmemSEXP, SEXP maxmemfracSEXP, SEXP maxtimeSEXP) {
 BEGIN_RCPP
     SEXP __sexp_result;
     {
         Rcpp::RNGScope __rngScope;
         Rcpp::traits::input_parameter< CharacterVector >::type passwd(passwdSEXP );
-        Rcpp::traits::input_parameter< size_t >::type maxmem(maxmemSEXP );
+        Rcpp::traits::input_parameter< uint32_t >::type maxmem(maxmemSEXP );
         Rcpp::traits::input_parameter< double >::type maxmemfrac(maxmemfracSEXP );
         Rcpp::traits::input_parameter< double >::type maxtime(maxtimeSEXP );
         CharacterVector __result = HashPassword(passwd, maxmem, maxmemfrac, maxtime);
